port and fix tests for room schedule
also make the content-rich test better on what it tests.
