+++ tests/gbe/test_reports.py
@@ -24,54 +24,6 @@ class TestReports(TestCase):
         self.client = Client()
         self.profile = ProfileFactory()
 
-    def test_room_schedule_fail(self):
-        '''room_schedule view should load for privileged users,
-           and fail for others
-        '''
-        current_conference = ConferenceFactory()
-        login_as(self.profile, self)
-        response = self.client.get(
-            reverse('room_schedule',
-                    urlconf='gbe.reporting.urls'))
-        self.assertEqual(response.status_code, 403)
-
-    def test_room_schedule_succeed(self):
-        '''room_schedule view should load for privileged users,
-           and fail for others
-        '''
-        Conference.objects.all().delete()
-        context = ClassContext()
-        one_day = timedelta(1)
-        ConferenceDayFactory(conference=context.conference,
-                             day=context.sched_event.starttime.date())
-        ConferenceDayFactory(
-            conference=context.conference,
-            day=context.sched_event.starttime.date()+one_day)
-        context.schedule_instance(
-            starttime=context.sched_event.starttime + one_day)
-        current_conference = context.conference
-        grant_privilege(self.profile, 'Act Reviewers')
-        login_as(self.profile, self)
-        response = self.client.get(
-            reverse('room_schedule',
-                    urlconf='gbe.reporting.urls',
-                    args=[context.room.pk]))
-        self.assertEqual(response.status_code, 200)
-
-    def test_room_schedule_by_conference(self):
-        '''room_schedule view should load for privileged users,
-           and fail for others
-        '''
-        Conference.objects.all().delete()
-        conf = ConferenceFactory()
-        grant_privilege(self.profile, 'Act Reviewers')
-        login_as(self.profile, self)
-        response = self.client.get(
-            reverse('room_schedule',
-                    urlconf='gbe.reporting.urls'),
-            data={'conf_slug': conf.conference_slug})
-        self.assertEqual(response.status_code, 200)
-
     def test_room_setup_not_visible_without_permission(self):
         '''room_setup view should load for privileged users,
            and fail for others
